fix degrader parsing; complete degrader test example

+++ parser/element.h
@@ -62,7 +62,7 @@ struct Element {
   double waveLength; ///< for laser wire and 3d transforms
   double gradient; ///< for rf cavities
   double phi, theta, psi; ///< for 3d transforms
-  double numberWedges; ///< for degrader
+  int numberWedges; ///< for degrader
   double wedgeHeight; ///< for degrader
   double degraderHeight; ///< for degrader
   double materialThickness; ///< for degrader
